requirements.txt
torch
torchvision
torchaudio
transformers
accelerate
datasets
peft
langchain
fastapi
uvicorn
python-dotenv
huggingface-hub